// lib/api/model/modelHabit.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Jeffries Homeapp API
 * Backend REST API for Jeffries Homeapp.
 * OpenAPI spec version: 1.0
 */

export interface ModelHabit {
  aangemaakt?: string;
  aangepaste_dagen?: number[];
  beschrijving?: string;
  doel_aantal?: number;
  doel_tijd?: string;
  doel_waarde?: number;
  eenheid?: string;
  emoji?: string;
  financie_categorie?: string;
  frequentie?: string;
  gepauzeer_om?: string;
  gewijzigd?: string;
  huidige_streak?: number;
  id?: string;
  is_actief?: boolean;
  is_kwantitatief?: boolean;
  is_pauze?: boolean;
  kleur?: string;
  langste_streak?: number;
  moeilijkheid?: string;
  naam?: string;
  rooster_filter?: string;
  totaal_voltooid?: number;
  totaal_xp?: number;
  type?: string;
  user_id?: string;
  volgorde?: number;
  xp_per_voltooiing?: number;
}