// lib/api/model/modelHabitBadge.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Jeffries Homeapp API
 * Backend REST API for Jeffries Homeapp.
 * OpenAPI spec version: 1.0
 */

export interface ModelHabitBadge {
  badge_id?: string;
  behaald_op?: string;
  beschrijving?: string;
  emoji?: string;
  habit_id?: string;
  id?: string;
  naam?: string;
  user_id?: string;
  xp_bonus?: number;
}